requirements.txt
appdirs==1.4.3
beautifulsoup4==4.4.0
bleach==1.5.0
click==6.7
defusedxml==0.5.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-annoying==0.10.3
django-bootstrap-form==3.2.1
django-bootstrap3==9.0.0
django-bower==5.2.0
django-cors-headers==1.1.0
django-debug-toolbar==1.5
django-extra-views==0.9.0
django-filer==1.2.7
django-filter==1.0.2
django-geojson==2.10.0
django-geoposition==0.3.0
django-google-maps==0.2.1
django-hero-slider==1.3.1
django-hvad==1.8.0
django-libs==1.67.12
django-mptt==0.8.7
django-polymorphic==1.0.2
django-sql-explorer==1.1.1
django-tables2==1.10.0
djangorestframework==3.6.3
easy-thumbnails==2.3
gunicorn==19.6.0
html5lib==0.9999999
icalendar==3.11.4
itsdangerous==0.24
Jinja2==2.9.6
jsonfield==1.0.3
Markdown==2.6.7
MarkupSafe==1.0
mock==2.0.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pbr==3.0.0
Pillow==4.0.0
psycopg2==2.6.2
py==1.4.33
PyJWT==1.5.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.6.0
python-decouple==2.3
python-firebase==1.2
python3-openid==3.1.0
pytz==2017.2
reportlab==3.4.0
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.3.0
sorl-thumbnail==12.4a1
sqlparse==0.2.3
static3==0.7.0
tox==1.7.2
twilio==6.0.0
Twisted==15.3.0
unicodecsv==0.14.1
Unidecode==0.4.20
Unipath==1.1
urllib3==1.21.1
virtualenv==15.1.0
vobject==0.9.5
Werkzeug==0.12.2
whitenoise==3.2.2
WTForms==2.1
zope.interface==4.4.0
django-easy-pdf==0.2.0
WeasyPrint>=0.34
